Update patient models to include patient fields
The current patient models only contain a few fields. I added the
remainder, and updated the character data types to use mostly text
instead of varchar

I updated the code for database transactions in GetPatientByID,
InsertPatient, DeletePatientByID, and tests accordingly

I also updated the id type to int32, since it is an integer type in
postgres

+++ domain/patient.go
@@ -17,17 +17,17 @@ type Patient struct {
 }
 
 type PatientUseCase interface {
-	GetPatientByID(ctx context.Context, id int64) (*Patient, error) // Takes an ID and returns a Patient object
-	DeletePatientByID(ctx context.Context, id int64) (int64, error) // Deletes a Patient by ID
-	UpdatePatientByID(ctx context.Context, id int64, patient *Patient) (int64, error)
-	InsertPatient(ctx context.Context, patient *Patient) (int64, error) // Creates a new patient and inserts in database
+	GetPatientByID(ctx context.Context, id int32) (*Patient, error) // Takes an ID and returns a Patient object
+	DeletePatientByID(ctx context.Context, id int32) (int32, error) // Deletes a Patient by ID
+	UpdatePatientByID(ctx context.Context, id int32, patient *Patient) (int32, error)
+	InsertPatient(ctx context.Context, patient *Patient) (int32, error) // Creates a new patient and inserts in database
 	GetAllFromAdmin(ctx context.Context) ([]models.Admin, error)
 }
 
 type PatientRepository interface {
-	GetPatientByID(ctx context.Context, id int64) (*Patient, error) // Takes an ID and returns a Patient object
-	DeletePatientByID(ctx context.Context, id int64) (int64, error) // Deletes a Patient by ID
-	UpdatePatientByID(ctx context.Context, id int64, patient *Patient) (int64, error)
-	InsertPatient(ctx context.Context, patient *Patient) (int64, error) // Creates a new patient and inserts in database
+	GetPatientByID(ctx context.Context, id int32) (*Patient, error) // Takes an ID and returns a Patient object
+	DeletePatientByID(ctx context.Context, id int32) (int32, error) // Deletes a Patient by ID
+	UpdatePatientByID(ctx context.Context, id int32, patient *Patient) (int32, error)
+	InsertPatient(ctx context.Context, patient *Patient) (int32, error) // Creates a new patient and inserts in database
 	GetAllFromAdmin(ctx context.Context) ([]models.Admin, error)        // returns all entries from a given category
 }
